Adjust config route
